refactor(npm-scripts): extract getWorkspaces() function
Because I am shortly going to need it to support our TypeScript build in
liferay-portal.

Test plan:

1.  Copy files into a `liferay-portal` checkout; ie.

        cp -R ~/code/liferay-frontend-projects/projects/npm-tools/packages/npm-scripts/src \
              ~modules/node_modules/@liferay/npm-scripts/

    Where `~modules` is a shortcut stored in directory hash table:

        https://github.com/wincent/wincent/blob/632aa515e22a/aspects/dotfiles/files/.zsh/hash#L24

2.  Confirm that all tests still run and pass; ie.

        cd ~modules
        portool test

    Which basically does:

        https://github.com/wincent/wincent/blob/632aa515e22a/aspects/dotfiles/files/.zsh/liferay/bin/portool-helpers/test#L13-L15
