+++ projects/npm-tools/packages/npm-scripts/src/utils/getJestModuleNameMapper.js
@@ -6,17 +6,13 @@
 const fs = require('fs');
 const path = require('path');
 
-const expandGlobs = require('./expandGlobs');
-const findRoot = require('./findRoot');
-const log = require('./log');
+const getWorkspaces = require('./getWorkspaces');
 
 /**
  * Where source really lives, relative to the directory containing the
  * "package.json" file.
  */
 const SRC_PATH = ['src', 'main', 'resources', 'META-INF', 'resources'];
-
-const IGNORE_GLOBS = ['node_modules/**'];
 
 /**
  * Returns a Jest "moduleNameMapper" configuration that enables tests to
@@ -33,10 +29,9 @@ const IGNORE_GLOBS = ['node_modules/**'];
  *
  * We create such mappings by:
  *
- *    1. Reading the Yarn workspace globs defined in the top-level
- *       "modules/package.json".
- *    2. Expanding those globs to identify possible projects.
- *    3. Selecting only projects which have a "package.json" with a "main"
+ *    1. Iterating over projects identified by the Yarn workspace globs
+ *       defined in the top-level "modules/package.json".
+ *    2. Selecting only projects which have a "package.json" with a "main"
  *       property that points to an existing file under
  *       "src/main/resources/META-INF/resources".
  *
@@ -51,89 +46,62 @@ function getJestModuleNameMapper() {
 	// from projects under "modules/apps" but not from those under
 	// "modules/private/apps".
 
-	const root = findRoot();
+	const cwd = process.cwd();
 
-	if (root) {
-		const cwd = process.cwd();
+	const mappings = {};
 
-		try {
-			process.chdir(root);
+	const projects = getWorkspaces();
 
-			const {workspaces} = JSON.parse(
-				fs.readFileSync('package.json', 'utf8')
-			);
+	projects.forEach((project) => {
+		const packageJson = path.join(project, 'package.json');
 
-			const mappings = {};
-			const projects = expandGlobs(workspaces.packages, IGNORE_GLOBS, {
-				maxDepth: 3,
-				type: 'directory',
-			});
+		const {main, name} = JSON.parse(fs.readFileSync(packageJson, 'utf8'));
 
-			projects.forEach((project) => {
-				const packageJson = path.join(project, 'package.json');
+		if (main) {
+			const entry = path.join(project, ...SRC_PATH, main);
 
-				if (fs.existsSync(packageJson)) {
-					const {main, name} = JSON.parse(
-						fs.readFileSync(packageJson, 'utf8')
+			// Handle typical formats for "main":
+			//
+			// - index        -> index.js
+			// - index.es     -> index.es.js
+			// - index.es.js  -> index.es.js
+			// - index.js     -> index.js
+			// - index.js     -> index.ts
+
+			const candidates = [entry];
+
+			if (entry.endsWith('.js')) {
+				candidates.push(entry.replace(/\.js$/, '.ts'));
+			}
+			else {
+				candidates.push(entry + '.js');
+			}
+
+			for (let i = 0; i < candidates.length; i++) {
+				const candidate = candidates[i];
+
+				if (fs.existsSync(candidate)) {
+					const resources = path.relative(
+						cwd,
+						path.join(project, ...SRC_PATH)
 					);
 
-					if (main) {
-						const entry = path.join(project, ...SRC_PATH, main);
+					mappings[`^${name}$`] = `<rootDir>${path.relative(
+						cwd,
+						candidate
+					)}`;
 
-						// Handle typical formats for "main":
-						//
-						// - index        -> index.js
-						// - index.es     -> index.es.js
-						// - index.es.js  -> index.es.js
-						// - index.js     -> index.js
-						// - index.js     -> index.ts
+					mappings[`^${name}/(.*)`] = `<rootDir>${resources}/$1`;
 
-						const candidates = [entry];
-
-						if (entry.endsWith('.js')) {
-							candidates.push(entry.replace(/\.js$/, '.ts'));
-						}
-						else {
-							candidates.push(entry + '.js');
-						}
-
-						for (let i = 0; i < candidates.length; i++) {
-							const candidate = candidates[i];
-
-							if (fs.existsSync(candidate)) {
-								const resources = path.relative(
-									cwd,
-									path.join(project, ...SRC_PATH)
-								);
-
-								mappings[
-									`^${name}$`
-								] = `<rootDir>${path.relative(cwd, candidate)}`;
-
-								mappings[
-									`^${name}/(.*)`
-								] = `<rootDir>${resources}/$1`;
-
-								break;
-							}
-						}
-					}
+					break;
 				}
-			});
+			}
+		}
+	});
 
-			return {
-				moduleNameMapper: mappings,
-			};
-		}
-		catch (error) {
-			log(`getJestModuleNameMapper(): error \`${error}\``);
-		}
-		finally {
-			process.chdir(cwd);
-		}
-	}
-
-	return {};
+	return {
+		moduleNameMapper: mappings,
+	};
 }
 
 module.exports = getJestModuleNameMapper;
